reanaming app to something less confusing
