--- knowledge_map/home/templates/knowledge_map/create_map.html
@@ -1,1 +1,0 @@
-{% extends "knowledge_map/base.html" %}
