--- knowledge_map/home/templates/knowledge_map/homepage.html
@@ -1,2 +1,0 @@
-{% load static %}
-{% extends "knowledge_map/base.html" %}
