--- knowledge_map/home/templates/knowledge_map/quiz.html
@@ -1,1 +1,0 @@
-{% extends "knowledge_map/base.html" %}
