+++ knowledge_map/knowledge_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class KnowledgeAppConfig(AppConfig):
+    name = 'knowledge_app'
