+++ knowledge_map/knowledge_app/templates/knowledge_app/homepage.html
@@ -1,0 +1,6 @@
+{% load static %}
+{% extends "knowledge_app/base.html" %}
+
+{% block content %}
+<h1>Welcome to the Homepage</h1>
+{% endblock %}
