+++ knowledge_map/knowledge_app/templates/knowledge_app/quiz.html
@@ -1,0 +1,1 @@
+{% extends "knowledge_app/base.html" %}
