+++ knowledge_map/knowledge_app/urls.py
@@ -3,7 +3,7 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name="index"),
-    path('', views.homepage, name='homepage'),
+    path('homepage/', views.homepage, name='homepage'),
     path('new_map/', views.create_map, name='create_map'),
     path('quiz/', views.quiz, name='quiz'),
     path('progress/', views.progress, name='progress'),
