+++ knowledge_map/knowledge_map/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'home',
+    'knowledge_app',
 ]
 
 MIDDLEWARE = [
