+++ knowledge_map/knowledge_map/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path("", include("home.urls")),
+    path("", include("knowledge_app.urls")),
 ]
